@@ -1,63 +1,25 @@
 {
   "name": "mcode-coverage-checker",
   "homepage": "http://mcode.github.io/mcode-coverage-checker",
-  "version": "0.1.0",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "contributors": [
+    "Alden Brown <ajbrown@mitre.org>",
     "Julian Carter <jacarter@mitre.org>",
     "Dylan Mendelowitz <dmendelowitz@mitre.org>",
     "Dylan Phelan <dtphelan1@gmail.com>"
   ],
-  "private": true,
-  "dependencies": {
-    "@emotion/react": "^11.10.4",
-    "@formkit/auto-animate": "^1.0.0-beta.3",
-    "@iconify/react": "^4.0.0",
-    "@table-library/react-table-library": "^4.0.19",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.0.0",
-    "fhirpath": "^2.12.0",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-circular-progressbar": "^2.1.0",
-    "react-dom": "^17.0.2",
-    "react-json-view": "^1.21.3",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.0",
-    "recharts": "^2.1.14",
-    "recoil": "^0.7.6",
-    "uuid": "^9.0.0",
-    "web-vitals": "^2.1.4"
-  },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "jest",
-    "eject": "react-scripts eject",
+    "lint": "eslint \"./**/*.js\"",
+    "lint-fix": "eslint \"./**/*.js\" --fix",
     "prepare": "husky install",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "lint": "eslint \"./**/*.js\"",
-    "lint-fix": "eslint \"./**/*.js\" --fix"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "jest"
   },
   "lint-staged": {
     "src/**/*.{js,jsx}": [
@@ -67,21 +29,39 @@
       "prettier --write"
     ]
   },
+  "dependencies": {
+    "@formkit/auto-animate": "^1.0.0-beta.5",
+    "@iconify/react": "^4.0.1",
+    "@textea/json-viewer": "^2.13.0",
+    "axios": "^1.2.3",
+    "fhirpath": "^3.3.1",
+    "react": "^18.2.0",
+    "react-circular-progressbar": "^2.1.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.7.0",
+    "recharts": "^2.3.2",
+    "recoil": "^0.7.6",
+    "uuid": "^9.0.0"
+  },
   "devDependencies": {
-    "autoprefixer": "^10.4.2",
-    "eslint": "^8.9.0",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "@vitejs/plugin-react": "^3.0.0",
+    "autoprefixer": "^10.4.13",
+    "eslint": "^8.32.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "gh-pages": "^3.2.3",
-    "husky": "^7.0.0",
-    "jest": "^27.5.1",
-    "lint-staged": "^12.3.4",
-    "postcss": "^8.4.6",
-    "prettier": "^2.5.1",
-    "tailwindcss": "^3.0.23"
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "gh-pages": "^5.0.0",
+    "husky": "^8.0.3",
+    "jest": "^29.3.1",
+    "lint-staged": "^13.1.0",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.3",
+    "tailwindcss": "^3.2.4",
+    "vite": "^4.0.0"
   }
 }
