@@ -1,14 +1,13 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 import App from './pages/App';
 import FileUploadPage from './pages/FileUploadPage';
-import reportWebVitals from './reportWebVitals';
 import PageWrapper from './components/PageWrapper';
 import './index.css';
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root')).render(
   <RecoilRoot>
     <Router basename="mcode-coverage-checker">
       <PageWrapper>
@@ -19,10 +18,4 @@ ReactDOM.render(
       </PageWrapper>
     </Router>
   </RecoilRoot>,
-  document.getElementById('root'),
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
