@@ -1,8 +1,8 @@
-const TumorMarkerTestVS = require('./valueSets/ValueSet-mcode-tumor-marker-test-vs.json');
-const SecondaryCancerDisorderVS = require('./valueSets/ValueSet-mcode-secondary-cancer-disorder-vs.json');
-const PrimaryCancerDisorderVS = require('./valueSets/ValueSet-mcode-primary-cancer-disorder-vs.json');
-const RadiotherapyVolumeTypeVS = require('./valueSets/ValueSet-mcode-radiotherapy-volume-type-vs.json');
-const GenomicSpecimenTypeVS = require('./valueSets/ValueSet-mcode-genomic-specimen-type-vs.json');
+import TumorMarkerTestVS from './valueSets/ValueSet-mcode-tumor-marker-test-vs.json';
+import SecondaryCancerDisorderVS from './valueSets/ValueSet-mcode-secondary-cancer-disorder-vs.json';
+import PrimaryCancerDisorderVS from './valueSets/ValueSet-mcode-primary-cancer-disorder-vs.json';
+import RadiotherapyVolumeTypeVS from './valueSets/ValueSet-mcode-radiotherapy-volume-type-vs.json';
+import GenomicSpecimenTypeVS from './valueSets/ValueSet-mcode-genomic-specimen-type-vs.json';
 
 // Map names used in VS to the corresponding JSON
 const DEFAULTVSLOOKUP = {
@@ -56,6 +56,4 @@ class ValueSetCodeChecker {
   }
 }
 
-module.exports = {
-  ValueSetCodeChecker,
-};
+export { ValueSetCodeChecker };
